feat(query): Raw and Periodic Samples query support
* Now able to issue LogicalPlan queries that extract raw data and periodic samples from ingested time series.
* Late materialization of result per ExecPlan into SerializableRangeVector

+++ query/src/main/scala/filodb/query/LogicalPlan.scala
@@ -60,7 +60,7 @@ case class PeriodicSeriesWithWindowing(rawSeries: RawSeries,
                                        start: Long,
                                        step: Int,
                                        end: Long,
-                                       window: Long,
+                                       window: Int,
                                        function: RangeFunctionId,
                                        functionArgs: Seq[Any] = Nil) extends PeriodicSeriesPlan
